Fix working dirs
