+++ commands/scanpullrequests_test.go
@@ -3,17 +3,17 @@ package commands
 import (
 	"context"
 	"fmt"
-	"github.com/jfrog/froggit-go/vcsutils"
-	"github.com/stretchr/testify/assert"
-	"path/filepath"
-	"testing"
-	"time"
-
 	"github.com/golang/mock/gomock"
 	"github.com/jfrog/frogbot/commands/testdata"
 	"github.com/jfrog/frogbot/commands/utils"
 	"github.com/jfrog/froggit-go/vcsclient"
+	coreconfig "github.com/jfrog/jfrog-cli-core/v2/utils/config"
 	"github.com/jfrog/jfrog-client-go/utils/io/fileutils"
+	"github.com/stretchr/testify/assert"
+	"os"
+	"path/filepath"
+	"strings"
+	"testing"
 )
 
 var gitParams = &utils.Repository{
@@ -37,69 +37,10 @@ type MockParams struct {
 }
 
 //go:generate go run github.com/golang/mock/mockgen@v1.6.0 -destination=testdata/vcsclientmock.go -package=testdata github.com/jfrog/froggit-go/vcsclient VcsClient
-func TestShouldScanPullRequestNewPR(t *testing.T) {
-	// Init mock
-	client := mockVcsClient(t)
-	prID := 0
-	client.EXPECT().ListPullRequestComments(context.Background(), gitParams.RepoOwner, gitParams.RepoName, prID).Return([]vcsclient.CommentInfo{}, nil)
-	// Run handleFrogbotLabel
-	shouldScan, err := shouldScanPullRequest(*gitParams, client, prID)
-	assert.NoError(t, err)
-	assert.True(t, shouldScan)
-}
-
-func TestShouldScanPullRequestReScan(t *testing.T) {
-	// Init mock
-	client := mockVcsClient(t)
-	prID := 0
-	client.EXPECT().ListPullRequestComments(context.Background(), gitParams.RepoOwner, gitParams.RepoName, prID).Return([]vcsclient.CommentInfo{
-		{Content: utils.GetSimplifiedTitle(utils.VulnerabilitiesPrBannerSource) + "text \n table\n text text text", Created: time.Unix(1, 0)},
-		{Content: utils.RescanRequestComment, Created: time.Unix(1, 1)},
-	}, nil)
-	shouldScan, err := shouldScanPullRequest(*gitParams, client, prID)
-	assert.NoError(t, err)
-	assert.True(t, shouldScan)
-}
-
-func TestShouldNotScanPullRequestReScan(t *testing.T) {
-	// Init mock
-	client := mockVcsClient(t)
-	prID := 0
-	client.EXPECT().ListPullRequestComments(context.Background(), gitParams.RepoOwner, gitParams.RepoName, prID).Return([]vcsclient.CommentInfo{
-		{Content: utils.GetSimplifiedTitle(utils.VulnerabilitiesPrBannerSource) + "text \n table\n text text text", Created: time.Unix(1, 0)},
-		{Content: utils.RescanRequestComment, Created: time.Unix(1, 1)},
-		{Content: utils.GetSimplifiedTitle(utils.NoVulnerabilityPrBannerSource) + "text \n table\n text text text", Created: time.Unix(3, 0)},
-	}, nil)
-	shouldScan, err := shouldScanPullRequest(*gitParams, client, prID)
-	assert.NoError(t, err)
-	assert.False(t, shouldScan)
-}
-
-func TestShouldNotScanPullRequest(t *testing.T) {
-	// Init mock
-	client := mockVcsClient(t)
-	prID := 0
-	client.EXPECT().ListPullRequestComments(context.Background(), gitParams.RepoOwner, gitParams.RepoName, prID).Return([]vcsclient.CommentInfo{
-		{Content: utils.GetSimplifiedTitle(utils.NoVulnerabilityPrBannerSource) + "text \n table\n text text text", Created: time.Unix(3, 0)},
-	}, nil)
-	shouldScan, err := shouldScanPullRequest(*gitParams, client, prID)
-	assert.NoError(t, err)
-	assert.False(t, shouldScan)
-}
 
 func mockVcsClient(t *testing.T) *testdata.MockVcsClient {
 	mockCtrl := gomock.NewController(t)
 	return testdata.NewMockVcsClient(mockCtrl)
-}
-
-func TestShouldNotScanPullRequestError(t *testing.T) {
-	// Init mock
-	client := mockVcsClient(t)
-	prID := 0
-	client.EXPECT().ListPullRequestComments(context.Background(), gitParams.RepoOwner, gitParams.RepoName, prID).Return([]vcsclient.CommentInfo{}, fmt.Errorf("Bad Request"))
-	shouldScan, err := shouldScanPullRequest(*gitParams, client, prID)
-	assert.Error(t, err)
-	assert.False(t, shouldScan)
 }
 
 func TestScanAllPullRequestsMultiRepo(t *testing.T) {
@@ -156,44 +97,6 @@ func TestScanAllPullRequestsMultiRepo(t *testing.T) {
 	expectedMessage = "<div align='center'>\n\n[![](https://raw.githubusercontent.com/jfrog/frogbot/master/resources/v2/noVulnerabilityBannerPR.png)](https://github.com/jfrog/frogbot#readme)\n\n</div>\n\n\n<div align=\"center\">\n\n[JFrog Frogbot](https://github.com/jfrog/frogbot#readme)\n\n</div>\n"
 	assert.Equal(t, expectedMessage, frogbotMessages[3])
 }
-
-func TestScanAllPullRequests(t *testing.T) {
-	// This integration test, requires JFrog platform connection details
-	server, restoreEnv := verifyEnv(t)
-	defer restoreEnv()
-	falseVal := false
-	gitParams.Git.GitProvider = vcsutils.BitbucketServer
-	params := utils.Params{
-		Scan: utils.Scan{
-			FailOnSecurityIssues: &falseVal,
-			Projects: []utils.Project{{
-				InstallCommandName: "npm",
-				InstallCommandArgs: []string{"i"},
-				WorkingDirs:        []string{"."},
-				UseWrapper:         &utils.TrueVal,
-			}},
-		},
-		Git: gitParams.Git,
-	}
-	repoParams := &utils.Repository{
-		OutputWriter: &utils.SimplifiedOutput{},
-		Server:       server,
-		Params:       params,
-	}
-	paramsAggregator := utils.RepoAggregator{}
-	paramsAggregator = append(paramsAggregator, *repoParams)
-	var frogbotMessages []string
-	client := getMockClient(t, &frogbotMessages, MockParams{repoParams.RepoName, repoParams.RepoOwner, "test-proj-with-vulnerability", "test-proj"})
-	scanAllPullRequestsCmd := &ScanAllPullRequestsCmd{}
-	err := scanAllPullRequestsCmd.Run(paramsAggregator, client)
-	assert.NoError(t, err)
-	assert.Len(t, frogbotMessages, 2)
-	expectedMessage := "**🚨 Frogbot scanned this pull request and found the below:**\n\n---\n## 📦 Vulnerable Dependencies\n---\n\n### ✍️ Summary \n\n| SEVERITY                | CONTEXTUAL ANALYSIS                  | DIRECT DEPENDENCIES                  | IMPACTED DEPENDENCY                   | FIXED VERSIONS                       |\n| :---------------------: | :----------------------------------: | :----------------------------------: | :-----------------------------------: | :---------------------------------: | \n| Critical | Not Applicable | minimist:1.2.5 | minimist:1.2.5 | [0.2.4], [1.2.6] |\n\n---\n### 👇 Details\n---\n\n\n#### minimist 1.2.5\n\n\n- **Severity** 💀 Critical\n- **Contextual Analysis:** Not Applicable\n- **Package Name:** minimist\n- **Current Version:** 1.2.5\n- **Fixed Versions:** [0.2.4],[1.2.6]\n- **CVE:** CVE-2021-44906\n\n**Description:**\n\n[Minimist](https://github.com/substack/minimist) is a simple and very popular argument parser. It is used by more than 14 million by Mar 2022. This package developers stopped developing it since April 2020 and its community released a [newer version](https://github.com/meszaros-lajos-gyorgy/minimist-lite) supported by the community.\n\n\nAn incomplete fix for [CVE-2020-7598](https://nvd.nist.gov/vuln/detail/CVE-2020-7598) partially blocked prototype pollution attacks. Researchers discovered that it does not check for constructor functions which means they can be overridden. This behavior can be triggered easily when using it insecurely (which is the common usage). For example:\n```\nvar argv = parse(['--_.concat.constructor.prototype.y', '123']);\nt.equal((function(){}).foo, undefined);\nt.equal(argv.y, undefined);\n```\nIn this example, `prototype.y`  is assigned with `123` which will be derived to every newly created object. \n\nThis vulnerability can be triggered when the attacker-controlled input is parsed using Minimist without any validation. As always with prototype pollution, the impact depends on the code that follows the attack, but denial of service is almost always guaranteed.\n\n**Remediation:**\n\n##### Development mitigations\n\nAdd the `Object.freeze(Object.prototype);` directive once at the beginning of your main JS source code file (ex. `index.js`), preferably after all your `require` directives. This will prevent any changes to the prototype object, thus completely negating prototype pollution attacks.\n\n\n\n\n\n[JFrog Frogbot](https://github.com/jfrog/frogbot#readme)"
-	assert.Equal(t, expectedMessage, frogbotMessages[0])
-	expectedMessage = "**👍 Frogbot scanned this pull request and found that it did not add vulnerable dependencies.** \n\n\n[JFrog Frogbot](https://github.com/jfrog/frogbot#readme)"
-	assert.Equal(t, expectedMessage, frogbotMessages[1])
-}
-
 func getMockClient(t *testing.T, frogbotMessages *[]string, mockParams ...MockParams) *testdata.MockVcsClient {
 	// Init mock
 	client := mockVcsClient(t)
@@ -228,4 +131,35 @@ func fakeRepoDownload(_ context.Context, _, _, testProject, targetDir string) er
 		return err
 	}
 	return fileutils.CopyDir(sourceDir, targetDir, true, []string{})
+}
+
+// Check connection details with JFrog instance.
+// Return a callback method that restores the credentials after the test is done.
+func verifyEnv(t *testing.T) (server coreconfig.ServerDetails, restoreFunc func()) {
+	url := strings.TrimSuffix(os.Getenv(utils.JFrogUrlEnv), "/")
+	username := os.Getenv(utils.JFrogUserEnv)
+	password := os.Getenv(utils.JFrogPasswordEnv)
+	token := os.Getenv(utils.JFrogTokenEnv)
+	if url == "" {
+		assert.FailNow(t, fmt.Sprintf("'%s' is not set", utils.JFrogUrlEnv))
+	}
+	if token == "" && (username == "" || password == "") {
+		assert.FailNow(t, fmt.Sprintf("'%s' or '%s' and '%s' are not set", utils.JFrogTokenEnv, utils.JFrogUserEnv, utils.JFrogPasswordEnv))
+	}
+	server.Url = url
+	server.XrayUrl = url + "/xray/"
+	server.ArtifactoryUrl = url + "/artifactory/"
+	server.User = username
+	server.Password = password
+	server.AccessToken = token
+	restoreFunc = func() {
+		utils.SetEnvAndAssert(t, map[string]string{
+			utils.JFrogUrlEnv:          url,
+			utils.JFrogTokenEnv:        token,
+			utils.JFrogUserEnv:         username,
+			utils.JFrogPasswordEnv:     password,
+			utils.GitAggregateFixesEnv: "FALSE",
+		})
+	}
+	return
 }
